// modules/loan/modules/b2b-request/enum/b2b-request-control-name.enum.ts
export enum AttributeB2BRequest {
  AMOUNT_TO_REQUEST = 'amount',
  FIXED_TERM = 'fixedTerm',
  WAY_PAY = 'paymentMethod',
  ACCOUNT_NUMBER = 'accountCharged',
  DUE_DATE = 'dueDate',
  INTEREST_FREQUENCY = 'interestPaymentFrequency',
  CAPITAL_FREQUENCY = 'capitalPaymentFrequency',
  ACCOUNT_CREDIT = 'accountAccredit',
}